fixed bug

+++ app/controllers/messenger_controller.rb
@@ -19,10 +19,11 @@ class MessengerController < ApplicationController
 
 			if @webhook["text"][0] == "done"
 				puts "INSIDE DONE"
-				@messages = Message.where(channel: @webhook["channel"][0], user_id: @webhook["user_id"][0])
-				puts @messages.inspect
-				@messages.each do |message|
-					message.destroy
+				@messages = Message.where(channel: @webhook["channel_name"][0], user_id: @webhook["user_id"][0])
+				if !@messages.empty?
+					@messages.each do |message|
+						message.destroy
+					end
 				end
 			end
